chore: add market code into margin mode
